+++ lib/crucible_signal/signal_ref.ex
@@ -63,6 +63,92 @@ defmodule CrucibleSignal.SignalRef do
     end
   end
 
+  def for_final_logits(attrs \\ []), do: typed_ref(:final_logits, attrs)
+
+  def for_layer_residual(layer_index, attrs \\ []) do
+    typed_ref(:middle_residuals, Keyword.put(normalize_keyword(attrs), :layer_index, layer_index))
+  end
+
+  def for_attention_map(layer_index, head_index, attrs \\ []) do
+    attrs =
+      attrs
+      |> normalize_keyword()
+      |> Keyword.put(:layer_index, layer_index)
+      |> Keyword.put(:head_index, head_index)
+
+    typed_ref(:attention_maps, attrs)
+  end
+
+  def for_mlp_gate(layer_index, attrs \\ []) do
+    typed_ref(:mlp_gates, Keyword.put(normalize_keyword(attrs), :layer_index, layer_index))
+  end
+
+  def for_norm_telemetry(layer_index, attrs \\ []) do
+    typed_ref(:norm_telemetry, Keyword.put(normalize_keyword(attrs), :layer_index, layer_index))
+  end
+
+  def for_moe_router(layer_index, attrs \\ []) do
+    typed_ref(
+      :moe_router_logits,
+      Keyword.put(normalize_keyword(attrs), :layer_index, layer_index)
+    )
+  end
+
+  def for_world_model(attrs \\ []), do: typed_ref(:world_model_state, attrs)
+  def for_verifier(attrs \\ []), do: typed_ref(:verifier_signal, attrs)
+
+  def for_logit_lens(layer_index, attrs \\ []) do
+    typed_ref(
+      :logit_lens_intermediate,
+      Keyword.put(normalize_keyword(attrs), :layer_index, layer_index)
+    )
+  end
+
+  def for_kv_cache(decode_step, attrs \\ []) do
+    typed_ref(:kv_cache_state, Keyword.put(normalize_keyword(attrs), :decode_step, decode_step))
+  end
+
+  def for_decoded_text(attrs \\ []), do: typed_ref(:decoded_text, attrs)
+
+  def validate!(%__MODULE__{} = ref) do
+    ref
+    |> Map.from_struct()
+    |> new!()
+  end
+
+  def validate!(attrs), do: new!(attrs)
+
+  defp typed_ref(signal_type, attrs) do
+    attrs =
+      attrs
+      |> normalize_keyword()
+      |> Keyword.put_new(:trace_id, "trace:unspecified")
+      |> Keyword.put_new(:signal_id, default_signal_id(signal_type, attrs))
+      |> Keyword.put_new(:signal_type, signal_type)
+      |> Keyword.put_new(:capture_mode, :summary)
+      |> Keyword.put_new(:redaction, :summary_only)
+
+    new!(attrs)
+  end
+
+  defp default_signal_id(signal_type, attrs) do
+    parts =
+      [
+        signal_type,
+        Keyword.get(attrs, :layer_index),
+        Keyword.get(attrs, :head_index),
+        Keyword.get(attrs, :token_index),
+        Keyword.get(attrs, :decode_step)
+      ]
+      |> Enum.reject(&is_nil/1)
+      |> Enum.map(&to_string/1)
+
+    Enum.join(parts, ":")
+  end
+
+  defp normalize_keyword(attrs) when is_map(attrs), do: Map.to_list(attrs)
+  defp normalize_keyword(attrs) when is_list(attrs), do: attrs
+
   defp normalize_attrs(attrs) when is_list(attrs), do: attrs |> Map.new() |> normalize_attrs()
 
   defp normalize_attrs(attrs) when is_map(attrs) do